Create AccessControlTabPanel and delete Groups/RolesTabPanel components
